requirements.txt
python-dotenv==0.10.3
paho_mqtt==1.4.0
